agenda_backend/pyproject.toml
[tool.poetry]
name = "agenda-backend"
version = "0.1.0"
description = ""
authors = ["Somesh Fengade <someshfengde@gmail.com>"]
readme = "README.md"
packages = [{include = "agenda_backend"}]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
groq = "^0.9.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
